ADM-627:[backend]fix: The story point when calculate and export from 0.5 to 0 in velocity and cycle time metrics

+++ backend/src/main/java/heartbeat/controller/board/dto/response/CardCollection.java
@@ -13,7 +13,7 @@ import java.util.List;
 @AllArgsConstructor
 public class CardCollection {
 
-	private int storyPointSum;
+	private double storyPointSum;
 
 	private int cardsNumber;
